t1_14/03

+++ src/scoreLabel.java
@@ -4,6 +4,7 @@ import java.awt.event.ActionListener;
 import java.awt.event.KeyAdapter;
 import java.awt.event.KeyEvent;
 import java.io.*;
+import java.util.Arrays;
 import java.util.HashMap;
 
 public class scoreLabel extends JPanel {
@@ -31,44 +32,97 @@ public class scoreLabel extends JPanel {
             @Override
             public void actionPerformed(ActionEvent e) {
 
-                saveScore();
+                try {
+                    saveScore();
+                } catch (IOException ex ) {
+                    throw new RuntimeException(ex);
+                } catch (ClassNotFoundException ex) {
+                    throw new RuntimeException(ex);
+                }
 
             }
         });
      }
-     void saveScore() {
-         try {
+     void saveScore() throws IOException, ClassNotFoundException {
+         if(file.exists()){
+
+
              FileInputStream fileInputStream = new FileInputStream("score_label.txt");
 
-             ObjectInputStream objectInputStream = new ObjectInputStream(fileInputStream);
-             HashMap hashMap = (HashMap)objectInputStream.readObject();
+             ObjectInputStream objectInputStream = null;
+             try {
+                 objectInputStream = new ObjectInputStream(fileInputStream);
+             } catch (IOException e) {
+                 throw new RuntimeException(e);
+             }
+             String[][] scoreTab = (String[][])objectInputStream.readObject();
+
+ /*
+            String[][] scoreTab = new String[10][2];
+             scoreTab[1][0]="maka";
+             scoreTab[1][1]="6";
+             scoreTab[2][0]="penn";
+             scoreTab[2][1]="3";
+             scoreTab[3][0]="siu";
+             scoreTab[3][1]="9";
+             scoreTab[4][0]="kut";
+             scoreTab[4][1]="1";
+             scoreTab[5][0]="dup";
+             scoreTab[5][1]="4";
+             scoreTab[6][0]="huu";
+             scoreTab[6][1]="8";
+             scoreTab[7][0]="na";
+             scoreTab[7][1]="6";
+             scoreTab[8][0]="Mareczek";
+             scoreTab[8][1]="34";
+             scoreTab[9][0]="Kent";
+             scoreTab[9][1]="123";
+             scoreTab[0][0]="James";
+             scoreTab[0][1]="1";
+*/
+             //scoreTab[1][1]= new String[]{{"James"},{"7"}};
+
+            for(int i=0;i<scoreTab.length;i++){
+
+                if(scoreTab[i][1] != null){
+                    scoreTab[i][0]=scoreField.getText()+" ";
+                    scoreTab[i][1]=String.valueOf(score);
+                    break;
+                }
+
+            }
 
 
-             fileInputStream.close();
-             objectInputStream.close();
-
-                hashMap.put(scoreField.getText(),score);
+             System.out.println("wyspisuje");
 
 
 
-
-             FileOutputStream fileOutputStream =new FileOutputStream("score_label.txt");
-
-             ObjectOutputStream objectOutputStream = new ObjectOutputStream(fileOutputStream);
-
-             objectOutputStream.writeObject(hashMap);
-             objectOutputStream.close();
-             fileOutputStream.close();
+             for(int i=0;i<scoreTab.length;i++){
+                 System.out.println();
+                 for(int j=0;j<scoreTab[i].length;j++){
+                     System.out.print(scoreTab[i][j]);
 
 
-         } catch (IOException | ClassNotFoundException e) {
-             throw new RuntimeException(e);
+                 }
+
+             }
+
+             try {
+                 FileOutputStream fileOutputStream = new FileOutputStream("score_label.txt");
+
+                 ObjectOutputStream objectOutputStream = new ObjectOutputStream(fileOutputStream);
+
+                objectOutputStream.writeObject(scoreTab);
+
+                 objectOutputStream.close();
+                 fileOutputStream.close();
+
+             } catch (IOException e) {
+                 throw new RuntimeException(e);
+             }
+
          }
 
      }
-}
-class hashTry implements Serializable{
-
-
 }
